Publish & Draft

+++ src/posts/views.py
@@ -3,14 +3,15 @@ from django.shortcuts import render,get_object_or_404,redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
 from .models  import Post
-from django.http import Http404
-from django.http import HttpResponse,HttpResponseRedirect
+from django.http import HttpResponse,HttpResponseRedirect,Http404
 from .forms import PostForm
 from django.contrib import messages
 
 
 def post_create(request):
-	form = PostForm(request.POST or None)
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	form = PostForm(request.POST or None,request.FILES or None)
 	if form.is_valid():
 		instance=form.save(commit=False)
 		print (form.cleaned_data.get("title"))
@@ -33,13 +34,14 @@ def post_create(request):
 
 
 
-def post_detail(request,id=None): #retrieve
+def post_detail(request,slug=None): #retrieve
 
-	instance = Post.objects.get(id=id)
+	instance = get_object_or_404(Post,slug=slug)
 
 	context = {
 		"title": instance.title,
 		"instance":instance,
+		# "image":instance.image,
 	}
 	return render(request, "post_detail.html", context)
 
@@ -73,12 +75,15 @@ def post_list(request): #list items
 	return render(request,"post_list.html",context)
 
 
-def post_update(request,id=None):
+def post_update(request,slug=None):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
 
-	instance = get_object_or_404(Post,id=id)
-	form = PostForm(request.POST or None, instance=instance)
+	instance = get_object_or_404(Post,slug=slug)
+	form = PostForm(request.POST or None, request.FILES or None,instance=instance)
 	if form.is_valid():
 		instance = form.save(commit=False)
+		instance.user = request.user
 		instance.save()
 		messages.success(request, "<a href='#'>Item</a> Saved", extra_tags='html_safe')
 		return HttpResponseRedirect(instance.get_absolute_url())
@@ -90,8 +95,10 @@ def post_update(request,id=None):
 	return render(request,"post_form.html",context)
 
 
-def post_delete(request,id=None):
-	instance = get_object_or_404(Post, id=id)
+def post_delete(request,slug=None):
+	if not request.user.is_staff or not request.user.is_superuser:
+		raise Http404
+	instance = get_object_or_404(Post, slug=slug)
 	instance.delete()
 	messages.success(request, "Successfully deleted")
 	return redirect("posts:list")
